+++ java/cuvs-java/src/main/java22/com/nvidia/cuvs/internal/common/Util.java
@@ -19,12 +19,10 @@ import static com.nvidia.cuvs.internal.common.LinkerHelper.C_CHAR;
 import static com.nvidia.cuvs.internal.common.LinkerHelper.C_FLOAT;
 import static com.nvidia.cuvs.internal.common.LinkerHelper.C_INT;
 import static com.nvidia.cuvs.internal.common.LinkerHelper.C_LONG;
-import static com.nvidia.cuvs.internal.common.LinkerHelper.C_POINTER;
 import static com.nvidia.cuvs.internal.panama.headers_h.cudaGetDeviceCount;
 import static com.nvidia.cuvs.internal.panama.headers_h.cudaGetDeviceProperties_v2;
 import static com.nvidia.cuvs.internal.panama.headers_h.cudaMemGetInfo;
 import static com.nvidia.cuvs.internal.panama.headers_h.cudaSetDevice;
-import static com.nvidia.cuvs.internal.panama.headers_h.cuvsRMMAlloc;
 import static com.nvidia.cuvs.internal.panama.headers_h.size_t;
 
 import com.nvidia.cuvs.GPUInfo;
@@ -350,17 +348,5 @@ public class Util {
     DLManagedTensor.dl_tensor(tensor, dlTensor);
 
     return tensor;
-  }
-
-  public static CloseableRMMAllocation allocateRMMSegment(long resourceHandle, long datasetBytes) {
-    try (var localArena = Arena.ofConfined()) {
-      MemorySegment datasetMemorySegment = localArena.allocate(C_POINTER);
-
-      var returnValue = cuvsRMMAlloc(resourceHandle, datasetMemorySegment, datasetBytes);
-      checkCuVSError(returnValue, "cuvsRMMAlloc");
-
-      var ret = datasetMemorySegment.get(C_POINTER, 0);
-      return new CloseableRMMAllocation(resourceHandle, datasetBytes, ret);
-    }
   }
 }
